Reorganize tests into full Django app
Needed for using things like apps.get_model().  Also allows for easier
reuse of test objects.
