+++ tests/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TestsConfig(AppConfig):
+    name = 'Tests'
+    label = 'tests'
