+++ tests/forms.py
@@ -1,0 +1,5 @@
+from django import forms
+
+
+class FooForm(forms.Form):
+    name = forms.CharField(max_length=5)
