+++ tests/models.py
@@ -1,0 +1,6 @@
+from django.db import models
+
+
+class FooModel(models.Model):
+    one = models.CharField(max_length=1)
+
